src/optimization_loop/requirements.txt
wandb
accelerate
ema-pytorch
tqd
einops
denoising-diffusion-pytorch
pymoo
cython